fix PR comments
