+++ web/client/themes/default/less/toc.less
@@ -147,10 +147,14 @@
                         margin-bottom: @padding-left-square-md;
                         .wfs-legend-icon {
                             margin-right: @padding-left-square-md;
+                            display: flex;
+                            align-items: center;
+                            
+                            svg {
+                                width: 16px;
+                                height: 16px;
+                            }
                         }
-                    }
-                    .wfs-legend-rule:last-child {
-                        margin-bottom: 0;
                     }
                 }
             }
